# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: MetanoiaDocs API
  version: '0.1'
  description: |
    Everything here is same-origin with the app. Two ways to authenticate:

    - **A session cookie** (`md_session`), which is what the browser uses.
    - **A personal access token** as `Authorization: Bearer <token>`, which is
      what everything else uses. Make one in Settings → API tokens. A token
      carries exactly the access of the person who made it — there are no
      scopes, because there is nothing a token should be able to do that its
      owner cannot.

    Times are ISO-8601. Dates on tasks are `YYYY-MM-DD` — a due date is a day,
    not an instant. Errors are `{ "error": "<sentence>" }` with a 4xx or 5xx
    status; the sentence is written to be shown to a person.
  license:
    name: MIT
    url: https://github.com/Xephyr-Labs/metanoiadocs/blob/main/LICENSE
servers:
  - url: http://localhost:8092
    description: A local instance from the quick start
  - url: https://docs.example.com
    description: Your instance
security:
  - bearerAuth: []
...
paths:
...
  /api/agent/runs/next:
    get:
      tags: [Agents]
      summary: Claim the next queued run
      description: |
        The runner's poll. Requires the token of an **agent** account — a
        person's token gets a 403 saying so, because a runner configured with
        the wrong token otherwise looks exactly like an empty queue.

        Claiming is atomic and skips locked rows, so several runners on one
        account each take a different run rather than queueing on a lock. A
        claimed run nobody reports on within `AGENT_RUN_STALE_MIN` minutes
        (default 30) goes back on the queue.
      responses:
        '200':
          description: A run, with everything needed to do it
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '204': { description: Nothing queued }
        '403': { $ref: '#/components/responses/Error' }
  /api/agent/runs/{id}/result:
    post:
      tags: [Agents]
      summary: Report what happened
      description: |
        Posts the result as a comment on the task's page, making that page if
        the task had none, and notifies whoever triggered the run.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ok: { type: boolean, default: true }
                result: { type: string, description: "Markdown, posted as the comment" }
                error: { type: [string, 'null'] }
                comment:
                  type: boolean
                  default: true
                  description: false to record the run without commenting.
      responses:
        '200':
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { const: true }
                  docId: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/Error' }
  /api/tasks/{id}/runs:
    get:
      tags: [Agents]
      summary: A task's recent runs
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Runs, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/RunRow' } }
    post:
      tags: [Agents]
      summary: Hand a task to an agent without assigning it
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentId]
              properties:
                agentId: { type: string }
                prompt: { type: string, description: Defaults to the task's title }
      responses:
        '200':
          description: The queued run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunRow' }
        '409': { description: That agent already has an open run on this task }
  /api/agent/runs/{id}/cancel:
    post:
      tags: [Agents]
      summary: Cancel a queued or running run
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The cancelled run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunRow' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token from Settings → API tokens.
    cookieAuth:
      type: apiKey
      in: cookie
      name: md_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    Error:
      description: A sentence saying what went wrong
      content:
        application/json:
          schema:
            type: object
            properties: { error: { type: string } }
  schemas:
...
    Run:
      type: object
      description: A claimed run, with its context read fresh at claim time.
      properties:
        id: { type: string }
        trigger: { type: string, enum: [assign, mention, manual] }
        prompt: { type: string, description: "What was asked — a comment's text, or the title" }
        createdAt: { type: string, format: date-time }
        task:
          type: [object, 'null']
          description: Null for a run on a page that is not a task's.
        page:
          type: [object, 'null']
          properties:
            id: { type: string }
            title: { type: string }
            text: { type: string }
        requestedBy: { type: [object, 'null'] }
    RunRow:
      type: object
      properties:
        id: { type: string }
        agent_id: { type: string }
        agent_name: { type: [string, 'null'] }
        task_id: { type: [string, 'null'] }
        doc_id: { type: [string, 'null'] }
        trigger: { type: string, enum: [assign, mention, manual] }
        status: { type: string, enum: [queued, running, done, failed, cancelled] }
        result: { type: string }
        error: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
...